+++ CHANGELOG.md
@@ -48,6 +48,7 @@ New features
 ------------
 
 * Functions defined in plugin modules can now be marked as `volatile` which prevents it from being optimized away even under `OptimizationLevel::Full`.
+* Added `Engine::max_functions` and `Engine::set_max_functions` to limit the maximum number of functions allowed in a script. This s to guard against DOS attacks -- e.g. a simple closure `||` (two characters) is a function. When `max_function` is exceeded during script compilation, a new parse error, `TooManyFunctions`, is returned.
 * `Engine::get_interned_string` is made public instead of gated under `internals`.
 
 Enhancements